+++ apps/ledger-live-desktop/src/config/urls.ts
@@ -18,6 +18,7 @@ const errors: Record<string, string> = {
   TronSendTrc20ToNewAccountForbidden: "https://support.ledger.com/article/6516823445533-zd",
   TronStakingDisable: "https://support.ledger.com/article/9949980566173-zd",
   OperatingSystemOutdated: "https://support.ledger.com/article/8083692639901-zd",
+  AddressesSanctionedError: "https://support.ledger.com/article/Why-Ledger-Complies-with-Sanctions",
 };
 
 export const urls = {
@@ -205,6 +206,9 @@ export const urls = {
     learnMore: "https://support.ledger.com/article/4409603715217-zd",
   },
   geoBlock: {
+    learnMore: "https://support.ledger.com/article/Why-Ledger-Complies-with-Sanctions",
+  },
+  sanctionCompliance: {
     learnMore: "https://support.ledger.com/article/Why-Ledger-Complies-with-Sanctions",
   },
 };