+++ apps/ledger-live-desktop/src/newArch/features/ModularDrawer/screens/FundAccount/components/ActionsContainer.tsx
@@ -18,9 +18,9 @@ import {
 const ActionItem = styled(Flex)`
   cursor: pointer;
   padding: 12px;
+
   border-radius: 8px;
   transition: background-color 0.2s ease;
-
   &:hover {
     background-color: ${p => p.theme.colors.opacityDefault.c05};
   }
@@ -31,9 +31,9 @@ const ActionItem = styled(Flex)`
 `;
 
 const Container = styled(Flex)`
-  width: calc(100% + 48px);
-  margin-left: -24px;
-  margin-right: -24px;
+  width: calc(100% + 72px);
+  margin-left: -36px;
+  margin-right: -36px;
 `;
 
 interface Props {
@@ -101,7 +101,7 @@ const ActionsContainer = ({ account, currencyId }: Props) => {
       height="100%"
       flex={1}
     >
-      <Flex flexDirection="column" width="100%" paddingX={24} paddingTop={24} rowGap={12}>
+      <Flex flexDirection="column" width="100%" paddingX={24} paddingTop={24}>
         {actions.map((action, index) => (
           <ActionItem
             key={index}
@@ -111,7 +111,7 @@ const ActionsContainer = ({ account, currencyId }: Props) => {
             columnGap={16}
           >
             <IconContainer icon={action.icon} />
-            <Flex flexDirection="column">
+            <Flex flexDirection="column" flex={1}>
               <Text fontSize={16} fontWeight="600">
                 {action.label}
               </Text>