+++ apps/ledger-live-desktop/src/newArch/features/ModularDrawer/screens/FundAccount/index.tsx
@@ -6,6 +6,7 @@ import { ActionsContainer } from "./components";
 import { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
 import { TrackAddAccountScreen } from "../../analytics/TrackAddAccountScreen";
 import { ADD_ACCOUNT_FLOW_NAME, ADD_ACCOUNT_PAGE_NAME } from "../../analytics/addAccount.types";
+import { Flex } from "@ledgerhq/react-ui/index";
 
 interface Props {
   account: Account;
@@ -23,7 +24,9 @@ const FundAccount = ({ account, currency, source }: Props) => {
         flow={ADD_ACCOUNT_FLOW_NAME}
         source={source}
       />
-      <AccountItem account={formattedAccount} />
+      <Flex width="100%" alignItems="center" marginBottom={24}>
+        <AccountItem account={formattedAccount} />
+      </Flex>
       <ActionsContainer account={account} currencyId={currency.id} />
     </>
   );