+++ apps/ledger-live-desktop/src/newArch/features/ModularDrawer/screens/ScanAccounts/components/Footer/index.tsx
@@ -1,20 +1,29 @@
 import { Button, Flex, Icons } from "@ledgerhq/react-ui";
-import React from "react";
-import { useTranslation } from "react-i18next";
-import useAddAccountAnalytics from "LLD/features/ModularDrawer/analytics/useAddAccountAnalytics";
+import { Account } from "@ledgerhq/types-live";
 import {
   ADD_ACCOUNT_EVENTS_NAME,
   ADD_ACCOUNT_FLOW_NAME,
 } from "LLD/features/ModularDrawer/analytics/addAccount.types";
+import useAddAccountAnalytics from "LLD/features/ModularDrawer/analytics/useAddAccountAnalytics";
+import React from "react";
+import { useTranslation } from "react-i18next";
+import { FOOTER_PADDING_BOTTOM_PX, FOOTER_PADDING_TOP_PX } from "../../../styles";
 
 export type FooterProps = {
+  handleConfirm: () => void;
+  importableAccounts: Account[];
   scanning: boolean;
   selectedIds: string[];
   stopSubscription: () => void;
-  handleConfirm: () => void;
 };
 
-export const Footer = ({ scanning, selectedIds, stopSubscription, handleConfirm }: FooterProps) => {
+export const Footer = ({
+  handleConfirm,
+  importableAccounts,
+  scanning,
+  selectedIds,
+  stopSubscription,
+}: FooterProps) => {
   const { t } = useTranslation();
 
   const { trackAddAccountEvent } = useAddAccountAnalytics();
@@ -29,19 +38,26 @@ export const Footer = ({ scanning, selectedIds, stopSubscription, handleConfirm 
   };
 
   return (
-    <Flex justifyContent="flex-end">
+    <Flex
+      justifyContent="flex-end"
+      paddingBottom={FOOTER_PADDING_BOTTOM_PX}
+      paddingTop={FOOTER_PADDING_TOP_PX}
+      zIndex={1}
+    >
       {scanning ? (
-        <Button
-          alignItems="center"
-          flex={1}
-          Icon={<Icons.Pause />}
-          iconPosition="left"
-          onClick={handleStopScanning}
-          size="xl"
-          variant="main"
-        >
-          {t("modularAssetDrawer.scanAccounts.cta.stopScanning")}
-        </Button>
+        importableAccounts.length === 0 ? null : (
+          <Button
+            alignItems="center"
+            flex={1}
+            Icon={<Icons.Pause />}
+            iconPosition="left"
+            onClick={handleStopScanning}
+            size="xl"
+            variant="main"
+          >
+            {t("modularAssetDrawer.scanAccounts.cta.stopScanning")}
+          </Button>
+        )
       ) : (
         <Button
           alignItems="center"