+++ apps/ledger-live-desktop/src/newArch/features/ModularDrawer/screens/ScanAccounts/index.tsx
@@ -1,19 +1,20 @@
 import { Box, Flex, Text } from "@ledgerhq/react-ui";
 import { AccountItem } from "@ledgerhq/react-ui/pre-ldls/components/AccountItem/AccountItem";
 import { Account } from "@ledgerhq/types-live";
-import { default as React } from "react";
-import { useTheme } from "styled-components";
 import { LoadingOverlay } from "LLD/components/LoadingOverlay";
-import { useScanAccounts, type UseScanAccountsProps } from "../../hooks/useScanAccounts";
-import { userThemeSelector } from "~/renderer/reducers/settings";
+import { TrackAddAccountScreen } from "LLD/features/ModularDrawer/analytics/TrackAddAccountScreen";
+import { default as React } from "react";
+import { useTranslation } from "react-i18next";
 import { useSelector } from "react-redux";
+import { useTheme } from "styled-components";
+import ErrorDisplay from "~/renderer/components/ErrorDisplay";
+import { userThemeSelector } from "~/renderer/reducers/settings";
+import { ADD_ACCOUNT_FLOW_NAME, ADD_ACCOUNT_PAGE_NAME } from "../../analytics/addAccount.types";
+import { ScrollContainer } from "../../components/ScrollContainer";
+import { useScanAccounts, type UseScanAccountsProps } from "../../hooks/useScanAccounts";
+import { CreatableAccountsList } from "./components/CreatableAccountsList";
 import { Footer } from "./components/Footer";
 import { ImportableAccountsList } from "./components/ImportableAccountsList";
-import { CreatableAccountsList } from "./components/CreatableAccountsList";
-import { useTranslation } from "react-i18next";
-import ErrorDisplay from "~/renderer/components/ErrorDisplay";
-import { TrackAddAccountScreen } from "LLD/features/ModularDrawer/analytics/TrackAddAccountScreen";
-import { ADD_ACCOUNT_FLOW_NAME, ADD_ACCOUNT_PAGE_NAME } from "../../analytics/addAccount.types";
 
 interface Props extends UseScanAccountsProps {
   analyticsPropertyFlow?: string;
@@ -89,7 +90,7 @@ const ScanAccounts = ({
         flow={ADD_ACCOUNT_FLOW_NAME}
       />
       {scanning ? <LoadingOverlay theme={currentTheme || "dark"} /> : null}
-      <Flex width="100%" alignItems="center">
+      <Flex marginBottom={24}>
         <Text
           fontSize={24}
           flex={1}
@@ -105,7 +106,7 @@ const ScanAccounts = ({
         </Text>
       </Flex>
 
-      <Flex flex={1} flexDirection="column" overflow="auto">
+      <ScrollContainer>
         {importableAccounts.length > 0 ? (
           <ImportableAccountsList
             scanning={scanning}
@@ -126,12 +127,13 @@ const ScanAccounts = ({
             renderAccount={renderAccount}
           />
         ) : null}
-      </Flex>
+      </ScrollContainer>
       <Footer
+        handleConfirm={handleConfirm}
+        importableAccounts={importableAccounts}
         scanning={scanning}
         selectedIds={selectedIds}
         stopSubscription={stopSubscription}
-        handleConfirm={handleConfirm}
       />
     </>
   );