+++ apps/ledger-live-mobile/src/utils/urls.tsx
@@ -80,6 +80,8 @@ export const urls = {
       "https://status.ledger.com?utm_source=ledger_live_mobile&utm_medium=self_referral&utm_content=error_apierrorwithmessage",
     StratisDown2021Warning: "https://support.ledger.com/article/115005175329-zd",
     TronSendTrc20ToNewAccountForbidden: "https://support.ledger.com/article/6516823445533-zd",
+    AddressesSanctionedError:
+      "https://support.ledger.com/article/Why-Ledger-Complies-with-Sanctions",
   },
   multipleAddresses: "https://support.ledger.com/article/360033802154-zd",
   delegation: