+++ e2e/mobile/page/wallet/transferMenu.drawer.ts
@@ -1,0 +1,12 @@
+export default class TransferMenuDrawer {
+  swapMenuButtonId = "swap-transfer-button";
+  transferButtonId = "transfer-button";
+
+  async open() {
+    await tapById(this.transferButtonId);
+  }
+
+  async navigateToSwap() {
+    await tapById(this.swapMenuButtonId);
+  }
+}