+++ libs/coin-modules/coin-xrp/.eslintrc.js
@@ -3,6 +3,8 @@ module.exports = {
     browser: true,
     es6: true,
   },
+  extends: ["plugin:import/typescript"],
+  plugins: ["import"],
   overrides: [
     {
       files: ["src/**/*.test.{ts,tsx}"],
@@ -16,5 +18,6 @@ module.exports = {
     "no-console": ["error", { allow: ["warn", "error"] }],
     "@typescript-eslint/no-empty-function": "off",
     "@typescript-eslint/no-explicit-any": "warn",
+    "import/no-cycle": ["error"],
   },
 };