+++ libs/ui/packages/react/src/pre-ldls/components/AccountItem/AccountItem.tsx
@@ -37,7 +37,7 @@ export type AccountItemProps = {
   backgroundColor?: string;
 };
 
-const Wrapper = styled.div<{ backgroundColor?: string }>`
+const Wrapper = styled.div<{ backgroundColor?: string; isClickable: boolean }>`
   ${withTokens(
     "spacing-xxxs",
     "spacing-xxs",
@@ -53,9 +53,7 @@ const Wrapper = styled.div<{ backgroundColor?: string }>`
   )}
 
   display: flex;
-  padding: var(--margin-s);
-  cursor: pointer;
-  border-radius: var(--radius-m);
+  cursor: ${p => (p.isClickable ? "pointer" : "default")};
   justify-content: space-between;
   align-items: center;
   width: 100%;
@@ -66,13 +64,20 @@ const Wrapper = styled.div<{ backgroundColor?: string }>`
 
   background-color: ${p => (p.backgroundColor ? p.backgroundColor : "transparent")};
 
-  :hover {
-    background-color: var(--colors-surface-transparent-hover);
-  }
+  ${p =>
+    p.isClickable
+      ? `
+    border-radius: var(--radius-m);
+    padding: var(--margin-s);
+    :hover {
+      background-color: var(--colors-surface-transparent-hover);
+    }
 
-  :active {
-    background-color: var(--colors-surface-transparent-pressed);
-  }
+    :active {
+      background-color: var(--colors-surface-transparent-pressed);
+    }
+  `
+      : ""}
 `;
 
 const ContentContainer = styled.div`
@@ -139,7 +144,7 @@ export const AccountItem = ({
   const { name, balance, fiatValue, protocol, address, ticker, cryptoId, parentId } = account;
 
   return (
-    <Wrapper onClick={onClick} backgroundColor={backgroundColor}>
+    <Wrapper onClick={onClick} backgroundColor={backgroundColor} isClickable={!!onClick}>
       <ContentContainer>
         <AccountInfoContainer>
           <NameRow>