+++ libs/ledgerjs/packages/hw-app-celo/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ledgerhq/hw-app-polkadot
 
+## 6.33.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-app-eth@6.45.10
+
+## 6.33.8-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-app-eth@6.45.10-next.0
+
 ## 6.33.7
 
 ### Patch Changes